fix(chatExtras): rename store to .ts
