+++ src/components/MessagesList/MessagesGroup/Message/Message.vue
@@ -123,7 +123,7 @@ import { CONVERSATION, MENTION, MESSAGE, PARTICIPANT } from '../../../../constan
 import { getTalkConfig, hasTalkFeature } from '../../../../services/CapabilitiesManager.ts'
 import { EventBus } from '../../../../services/EventBus.ts'
 import { useActorStore } from '../../../../stores/actor.ts'
-import { useChatExtrasStore } from '../../../../stores/chatExtras.js'
+import { useChatExtrasStore } from '../../../../stores/chatExtras.ts'
 import { getItemTypeFromMessage } from '../../../../utils/getItemTypeFromMessage.ts'
 
 const canSummarizeChat = hasTalkFeature('local', 'chat-summary-api')
