+++ src/components/Quote.vue
@@ -19,7 +19,7 @@ import { useMessageInfo } from '../composables/useMessageInfo.ts'
 import { AVATAR } from '../constants.ts'
 import { EventBus } from '../services/EventBus.ts'
 import { useActorStore } from '../stores/actor.ts'
-import { useChatExtrasStore } from '../stores/chatExtras.js'
+import { useChatExtrasStore } from '../stores/chatExtras.ts'
 import { getMessageIcon } from '../utils/getMessageIcon.ts'
 
 type DeletedParentMessage = Pick<ChatMessage, 'id' | 'deleted'>
