+++ src/composables/useTemporaryMessage.ts
@@ -8,7 +8,7 @@ import type { PrepareTemporaryMessagePayload } from '../utils/prepareTemporaryMe
 
 import { useStore } from 'vuex'
 import { useActorStore } from '../stores/actor.ts'
-import { useChatExtrasStore } from '../stores/chatExtras.js'
+import { useChatExtrasStore } from '../stores/chatExtras.ts'
 import { prepareTemporaryMessage } from '../utils/prepareTemporaryMessage.ts'
 
 /**
